H.265: Fix unintentional integer truncation
Rec. ITU-T H.265 5.2 Arithmetic operators

        / Integer division with truncation of the result toward zero.
        ÷ Used to denote division in mathematical equations where no
                truncation or rounding is intended

7.4.3.2.1 General sequence parameter set RBSP semantics

        PicWidthInCtbsY = Ceil( pic_width_in_luma_samples ÷ CtbSizeY ) (7-15)
        PicHeightInCtbsY = Ceil( pic_height_in_luma_samples ÷ CtbSizeY ) (7-17)

As this indicates that no truncation or rounding is intended, cast to double
first to avoid integer truncation.

Coverity CID 1450796

+++ epan/dissectors/packet-h265.c
@@ -1874,9 +1874,9 @@ dissect_h265_slice_segment_header(proto_tree *tree, tvbuff_t *tvb, packet_info *
 	unsigned MinCbLog2SizeY = log2_min_luma_coding_block_size_minus3 + 3;
 	unsigned CtbLog2SizeY = MinCbLog2SizeY + log2_diff_max_min_luma_coding_block_size;
 	unsigned CtbSizeY = 1 << CtbLog2SizeY;
-	double PicWidthInCtbsY = ceil(pic_width_in_luma_samples / CtbSizeY);
-        double PicHeightInCtbsY = ceil(pic_height_in_luma_samples / CtbSizeY);
-        double PicSizeInCtbsY = PicWidthInCtbsY * PicHeightInCtbsY;
+	double PicWidthInCtbsY = ceil((double)pic_width_in_luma_samples / CtbSizeY);
+	double PicHeightInCtbsY = ceil((double)pic_height_in_luma_samples / CtbSizeY);
+	double PicSizeInCtbsY = PicWidthInCtbsY * PicHeightInCtbsY;
 	unsigned nBits = (unsigned)(ceil(log2(PicSizeInCtbsY)));
 	unsigned i;
